+++ pages/api/replicache-push.ts
@@ -8,6 +8,7 @@ import {
 } from "../../backend/data";
 import type { NextApiRequest, NextApiResponse } from "next";
 import { ReplicacheTransaction } from "../../backend/replicache-transaction";
+import { getPullOrder } from "../../backend/pull-order";
 import { mutators } from "../../frontend/mutators";
 import { z } from "zod";
 import { jsonSchema } from "../../util/json";
@@ -51,7 +52,8 @@ const push = async (req: NextApiRequest, res: NextApiResponse) => {
       executor,
       spaceID,
       push.clientID,
-      nextVersion
+      nextVersion,
+      getPullOrder
     );
 
     for (let i = 0; i < push.mutations.length; i++) {
